Card Problem
You are given a deck containing N cards.  While holding the deck:

1. Take the top card off the deck and set it on the table
2. Take the next card off the top and put it on the bottom of the deck
in your hand.
3. Continue steps 1 and 2 until all cards are on the table.  This is a
round.
4. Pick up the deck from the table and repeat steps 1-3 until the deck
is in the original order.

Write a program to determine how many rounds it will take to put a
deck back into the original order.  This will involve creating a data
structure to represent the order of the cards. Do not use an array.
This program should be written in C only. It should take a number of
cards in the deck as a command line argument and write the result to
stdout.